Move window and split-panel sources into folders
App/ still kept the window-level sources and SplitPanel flat at the
project root next to the entry point. Finish the layout started with
App/Terminal:

  App/Windows/      MainWindow, MainWindows, WindowCloseGate, TransparentBackdrop
  App/Tabs/Panes/   SplitPanel (it owns the pane Tree, so it lives with it)

Includes take the folder path; the generated .g.h / .g.cpp includes
stay root-relative because cppwinrt writes them at the root of
Generated Files, and $(GeneratedFilesDir)Windows joins the include
path so MainWindow.g.h finds the XAML compiler's mirrored
MainWindow.xaml.g.h. Filters and DependentUpon follow the same rules
as the Terminal move.
