+++ App/Tabs/Panes/SplitPanel.idl
