+++ App/Windows/MainWindow.xaml.cpp
@@ -1,9 +1,9 @@
 #include "pch.h"
-#include "MainWindow.xaml.h"
+#include "Windows/MainWindow.xaml.h"
 #include "App.xaml.h"
 #include "Ghostty/CallbackDispatcher.h"
 #include "Ghostty/Config.h"
-#include "TransparentBackdrop.h"
+#include "Windows/TransparentBackdrop.h"
 #include "Host/KeyModifiers.h"
 #include "Interop/Encoding.h"
 #include "Display/PhysicalPixels.h"
