+++ App/Windows/WindowCloseGate.h
